separate files into folders
